% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_MR.R
\name{crosstab_MR}
\alias{crosstab_MR}
\title{A function to get crosstab of x by y in MapReduce manner.}
\usage{
crosstab_MR(dfs_data, x, y, ylevels)
}
\arguments{
\item{dfs_data}{A data frame in HDFS}

\item{x}{Column name X in the data frame}

\item{y}{Column name Y in the data frame}

\item{ylevels}{A char vector with all level names of variable/column Y}
}
\description{
Given a data frame with x and y variables, this MapReduce
function creates a crosstab of x by y.
The level names of y is required to label all rows and columns.
}
\examples{
from.dfs(crosstab_MR("/home/yihuang/mtcars.RData",x = 'am', y = 'gear', ylevels = c(3,4,5)));
}
\author{
Yihuang Kang
}

