% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadFestival.R
\docType{data}
\name{DownloadFestival}
\alias{DownloadFestival}
\title{"Download Festival" Data}
\format{data.frame}
\source{
http://www.discoveringstatistics.com/docs/DownloadFestival(No%20Outlier).dat
}
\usage{
data(DownloadFestival)
}
\description{
"Download Festival" is a rock and heavy metal festival in the UK. This dataset contains 
people’s hygiene score over the three days of the festival using a standardized technique 
that results in a score ranging between 0 (you smell like a rotting corpse that’s hiding up 
a skunk’s anus) and 4 (you smell of sweet roses on a fresh spring day).
See the text p. 170 of "Discovering Statistics using R" by ANDY FIELD | JEREMY MILES | ZOË FIELD
}
\examples{
data(DownloadFestival)
}
\references{
"Discovering Statistics using R" by ANDY FIELD | JEREMY MILES | ZOË FIELD,  p. 170
}
\keyword{datasets}

