% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2char.R
\name{factor2char}
\alias{factor2char}
\title{To convert factor variable into character variable}
\usage{
factor2char(factorVar)
}
\arguments{
\item{factorVar}{A factor variable to convert.}
}
\description{
To convert factor variable into character variable.
}
\examples{
factor2char(CO2$Type)
}
\author{
Yihuang Kang
}

