% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoker.R
\docType{data}
\name{smoker}
\alias{smoker}
\title{356 people's smoking status with their socioeconomic status (SES)}
\format{data.frame}
\source{
http://www.cyclismo.org/tutorial/R/_static/smoker.csv
}
\usage{
data(smoker)
}
\description{
356 people's smoking status with their socioeconomic status
For each person it was determined whether or not they are current smokers, former smokers, or have never smoked.
Also, for each person their socioeconomic status was determined (low, middle, or high).
}
\examples{
data(smoker)
}
\references{
Moore, McCabe's, Introduction to the Practice of Statistics, 4th edition, p. 629
}
\keyword{datasets}

